+++ Game.Core/Application/Abstraction/IUseCase.cs
@@ -1,0 +1,6 @@
+namespace Game.Core.Application.Abstraction;
+
+public interface IUseCase<in TRequest>
+{
+    public Result Execute(TRequest request);
+}